separate track_service to multiple files
