+++ src/states.rs
@@ -2,7 +2,7 @@ use crate::error::TGVError;
 use crate::repository::Repository;
 use crate::repository::{AlignmentRepository, AlignmentRepositoryEnum};
 use crate::settings::Settings;
-use crate::track_service::{TrackCache, TrackService};
+use crate::tracks::{TrackCache, TrackService};
 use crate::traits::GenomeInterval;
 use crate::{
     alignment::Alignment,
